+++ src/models/transaction.rs
@@ -8,7 +8,7 @@ pub enum ActivityType {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct TransactionRequest {
-    pub user_id: u64,
+    pub user_id: i32,
     pub amount: f64,
     pub activity: String,
     pub activity_type: ActivityType,
